+++ src/main/java/com/nguyenquyhy/discordbridge/commands/ReloadCommand.java
@@ -4,12 +4,14 @@ import com.nguyenquyhy.discordbridge.DiscordBridge;
 import com.nguyenquyhy.discordbridge.logics.ConfigHandler;
 import com.nguyenquyhy.discordbridge.logics.LoginHandler;
 import com.nguyenquyhy.discordbridge.models.GlobalConfig;
+import de.btobastian.javacord.DiscordAPI;
 import org.slf4j.Logger;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.spec.CommandExecutor;
+import org.spongepowered.api.text.Text;
 
 /**
  * Created by Hy on 1/5/2016.
@@ -22,7 +24,8 @@ public class ReloadCommand implements CommandExecutor {
             GlobalConfig config = ConfigHandler.loadConfiguration();
             DiscordBridge.getInstance().setConfig(config);
             logger.info("Configuration reloaded!");
-            LoginHandler.loginBotAccount();
+            commandSource.sendMessage(Text.of("Configuration reloaded!"));
+
             return CommandResult.success();
         } catch (Exception e) {
             logger.error("Cannot reload configuration!", e);